refactor: Extract test helpers from test file
